Remove runtime dependence on policycoreutils from the reference policy.
The only thing refpol should depend on at runtime are the things
necessary to load the policy.  If init is patched then we only need
the config.  It may be necessary to include policycoreutil-loadpolicy
as well in the absence of a patched init.
(cherry picked from commit c20a470b65f2c6dbc33069b10a9a302794ac30f5)

Signed-off-by: Philip Tricca <flihp@twobit.us>